Models added and bug fixed

+++ src/main/java/com/nusan/nusanapi/repository/PagCliRepository.java
@@ -3,5 +3,5 @@ package com.nusan.nusanapi.repository;
 import com.nusan.nusanapi.model.Client;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface PagCliRepository extends PagingAndSortingRepository<Client, Long> {
+public interface PagCliRepository extends PagingAndSortingRepository<Client, Integer> {
 }
